refactor: extract change-token form into a global ChangeTokenModal
Extract the inline access-token change form from SettingsView into a
standalone ChangeTokenModal component backed by a useChangeTokenModal
composable. The modal is rendered once in MainLayout so every page can
trigger it via openChangeTokenModal().
